updated the footer component
Include the Jest test library to carryout the unit tests.

created a Geolocate function

During the test, discovered other issues in the urls, utils, and operations file and refactored them accordingly to solve the issues
