+++ website/.babelrc
@@ -1,6 +1,4 @@
 {
-    "presets": ["@babel/preset-env", "@babel/preset-react"],
-    "plugins": [
-        ["@babel/plugin-transform-runtime"]
-    ]
+  "presets": ["@babel/preset-env", ["@babel/preset-react", { "runtime": "automatic" }]],
+  "plugins": [["@babel/plugin-transform-runtime"]]
 }
