+++ website/frontend/config/utils.js
@@ -1,13 +1,18 @@
 const URL_PATTERN = new RegExp(
-  '^(https?:\\/\\/)?' // protocol
-  + '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.?)+[a-z]{2,}|' // domain name
-  + '((\\d{1,3}\\.){3}\\d{1,3}))' // ip (v4) address
-  + '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' // port
-  + '(\\?[;&amp;a-z\\d%_.~+=-]*)?' // query string
-    + '(\\#[-a-z\\d_]*)?$',
-  'i',
+  '^(https?:\\/\\/)?' + // protocol
+    '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.?)+[a-z]{2,}|' + // domain name
+    '((\\d{1,3}\\.){3}\\d{1,3}))' + // ip (v4) address
+    '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // port
+    '(\\?[;&amp;a-z\\d%_.~+=-]*)?' + // query string
+    '(\\#[-a-z\\d_]*)?$',
+  'i'
 );
 
 export const validateUrl = (url) => URL_PATTERN.test(url);
 
-export const stripTrailingSlash = (url) => url.replace(/\/$/, '');
+export const stripTrailingSlash = (url) => {
+  if (typeof url === 'undefined') {
+    return url;
+  }
+  return url.replace(/\/$/, '');
+};
