+++ website/frontend/reduxStore/AirQlouds/operations.js
@@ -1,33 +1,32 @@
 import { isEmpty } from 'underscore';
-import { getAirQloudSummaryApi } from 'apis';
+import { getAirQloudSummaryApi } from '../../apis';
 import {
   LOAD_AIRQLOUDS_SUMMARY_SUCCESS,
   LOAD_AIRQLOUDS_SUMMARY_FAILURE,
   SET_CURRENT_AIRQLOUD_SUCCESS,
-  SET_CURRENT_AIRQLOUD_FAILURE,
+  SET_CURRENT_AIRQLOUD_FAILURE
 } from './actions';
 import { transformArray } from '../utils';
 
 export const setCurrentAirQloudData = (name) => (dispatch) => {
-  dispatch(
-    {
-      type: SET_CURRENT_AIRQLOUD_SUCCESS,
-      payload: name,
-    },
-  );
+  dispatch({
+    type: SET_CURRENT_AIRQLOUD_SUCCESS,
+    payload: name
+  });
 };
 
-export const loadAirQloudSummaryData = () => async (dispatch) => await getAirQloudSummaryApi()
-  .then((resData) => {
-    if (isEmpty(resData.airqlouds || [])) return;
-    dispatch({
-      type: LOAD_AIRQLOUDS_SUMMARY_SUCCESS,
-      payload: transformArray(resData.airqlouds, 'long_name'),
+export const loadAirQloudSummaryData = () => async (dispatch) =>
+  await getAirQloudSummaryApi()
+    .then((resData) => {
+      if (isEmpty(resData.airqlouds || [])) return;
+      dispatch({
+        type: LOAD_AIRQLOUDS_SUMMARY_SUCCESS,
+        payload: transformArray(resData.airqlouds, 'long_name')
+      });
+    })
+    .catch((err) => {
+      dispatch({
+        type: LOAD_AIRQLOUDS_SUMMARY_FAILURE,
+        payload: err && err.message
+      });
     });
-  })
-  .catch((err) => {
-    dispatch({
-      type: LOAD_AIRQLOUDS_SUMMARY_FAILURE,
-      payload: err && err.message,
-    });
-  });
