+++ website/frontend/src/components/__tests__/Geolocate.test.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import UserCountry from '../Geolocate';
+import { Provider } from 'react-redux';
+import configureStore from 'redux-mock-store';
+
+const mockStore = configureStore([]);
+
+describe('UserCountry', () => {
+  let store;
+  beforeEach(() => {
+    store = mockStore({
+      airQlouds: {
+        currentAirQloudData: null
+      }
+    });
+  });
+
+  it('renders without crashing', () => {
+    render(
+      <Provider store={store}>
+        <UserCountry
+          countries={[
+            {
+              name: 'Uganda'
+            }
+          ]}
+        />
+      </Provider>
+    );
+  });
+});
