+++ website/package.json
@@ -1,67 +1,77 @@
 {
-    "name": "website",
-    "version": "1.0.0",
-    "description": "AirQo website",
-    "main": "index.js",
-    "scripts": {
-        "serve": "webpack serve --config webpack.dev.config.js --mode development",
-        "standalone": "webpack serve --config webpack.fullapp.config.js --mode development",
-        "build": "webpack --mode production"
-    },
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "@babel/runtime": "^7.20.13",
-        "@emotion/react": "^11.10.5",
-        "@emotion/styled": "^11.10.5",
-        "@mui/icons-material": "^5.11.0",
-        "@mui/material": "^5.11.8",
-        "@reduxjs/toolkit": "^1.9.3",
-        "axios": "^1.3.2",
-        "date-fns": "^2.29.3",
-        "react": "^18.2.0",
-        "react-device-detect": "^2.2.3",
-        "react-dom": "^18.2.0",
-        "react-helmet": "^6.1.0",
-        "react-redux": "^8.0.5",
-        "react-router-dom": "^6.8.1",
-        "react-scroll": "^1.8.9",
-        "redux": "^4.2.1",
-        "redux-devtools-extension": "^2.13.9",
-        "redux-thunk": "^2.4.2",
-        "underscore": "^1.13.6"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.20.12",
-        "@babel/plugin-transform-runtime": "^7.19.6",
-        "@babel/preset-env": "^7.20.2",
-        "@babel/preset-react": "^7.18.6",
-        "@svgr/webpack": "^6.5.1",
-        "babel-loader": "^9.1.2",
-        "classnames": "^2.3.2",
-        "css-loader": "^6.7.3",
-        "dotenv": "^16.0.3",
-        "eslint": "^8.34.0",
-        "eslint-config-airbnb": "^19.0.4",
-        "eslint-plugin-import": "^2.27.5",
-        "eslint-plugin-jsx-a11y": "^6.7.1",
-        "eslint-plugin-react": "^7.32.2",
-        "eslint-plugin-react-hooks": "^4.6.0",
-        "file-loader": "^6.2.0",
-        "html-webpack-plugin": "^5.5.0",
-        "node-sass": "^8.0.0",
-        "postcss": "^8.4.21",
-        "postcss-loader": "^7.0.2",
-        "postcss-preset-env": "^8.0.1",
-        "prettier": "^2.8.4",
-        "sass-loader": "^13.2.0",
-        "style-loader": "^3.3.1",
-        "url-loader": "^4.1.1",
-        "webpack": "^5.75.0",
-        "webpack-cli": "^5.0.1",
-        "webpack-dev-server": "^4.11.1"
-    },
-    "engines": {
-        "node": ">=14.x <=18.x"
-    }
+  "name": "website",
+  "version": "1.0.0",
+  "description": "AirQo website",
+  "main": "index.js",
+  "scripts": {
+    "serve": "webpack serve --config webpack.dev.config.js --mode development",
+    "standalone": "webpack serve --config webpack.fullapp.config.js --mode development",
+    "build": "webpack --mode production",
+    "test": "jest --watchAll"
+  },
+  "jest": {
+    "testEnvironment": "jest-environment-jsdom"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@babel/runtime": "^7.20.13",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.8",
+    "@reduxjs/toolkit": "^1.9.3",
+    "axios": "^1.3.2",
+    "date-fns": "^2.29.3",
+    "react": "^18.2.0",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.1",
+    "react-scroll": "^1.8.9",
+    "redux": "^4.2.1",
+    "redux-devtools-extension": "^2.13.9",
+    "redux-thunk": "^2.4.2",
+    "underscore": "^1.13.6"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/plugin-transform-runtime": "^7.19.6",
+    "@babel/preset-env": "^7.22.9",
+    "@babel/preset-react": "^7.22.5",
+    "@svgr/webpack": "^6.5.1",
+    "@testing-library/react": "^14.0.0",
+    "babel-jest": "^29.6.2",
+    "babel-loader": "^9.1.2",
+    "classnames": "^2.3.2",
+    "css-loader": "^6.7.3",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.34.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.6.2",
+    "jest-environment-jsdom": "^29.6.2",
+    "node-sass": "^8.0.0",
+    "postcss": "^8.4.21",
+    "postcss-loader": "^7.0.2",
+    "postcss-preset-env": "^8.0.1",
+    "prettier": "^2.8.4",
+    "react-test-renderer": "^18.2.0",
+    "redux-mock-store": "^1.5.4",
+    "sass-loader": "^13.2.0",
+    "style-loader": "^3.3.1",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
+  },
+  "engines": {
+    "node": ">=14.x <=18.x"
+  }
 }
